Create .ruby-version
Per Netify The build is failing because the required Ruby version 2.6.2 is not installed on the build environment. Solution: To fix this issue, you need to ensure that Ruby 2.6.2 is installed in the build environment.   To do this Set a RUBY_VERSION environment variable.
Add a .ruby-version file to the site’s base directory in your repository. This file should contain the version number only, like 2.6.2, with no trailing newline.
If the version you select is preinstalled in your site’s selected build image, it will be available immediately. If not, your selected version will be installed using rvm and then cached to speed up subsequent builds.

If your build still fails after setting the Ruby version, it might be due to missing software required for the build command. You can tell the build system to install the software you need by including the proper configuration file, like a Gemfile for Ruby programs. The build system looks for the configuration file in your site’s base directory, which is the root of your repository by default. If your configuration file is located in a subdirectory, you will need to change your folder structure or set the site’s package directory to specify the location.